requirements.txt
Django~=2.1.4
wiki~=0.4.2
djangorestframework~=3.9.0
